优化员工业务代码

+++ sky-server/src/main/java/com/sky/mapper/EmployeeMapper.java
@@ -25,5 +25,5 @@ public interface EmployeeMapper {
 
     void update(Employee employee);
 
-    Employee getById(Integer id);
+    Employee getById(Long id);
 }
